kernel/os: Create common header with os_arch definitions
There are a lot of c&p in os_arch.h headers, including common
definitions and APIs so they can be extracted to common file to make
maintaining these headers easier (no need to sync changes everywhere).

Definitions can be still overwritten in arch header if required while
default values are set so most of archs do not to do this.
